shouldve done this before

+++ WebScrapers/OnSite/mangaReaderDotOrg.py
@@ -96,9 +96,10 @@ def print_links_in_reverse_order(url):
                 typeFile.write(type_value if type_value else '')
         
         id_path = os.path.join(directory_path, 'id.txt')
-        print(f"ID Path: {id_path}")
-        with open(id_path, 'w') as id_file:
-            id_file.write(count_directories(DIRECTORY_PATH))
+        if not os.path.exists(id_path):
+            print(f"ID Path: {id_path}")
+            with open(id_path, 'w') as id_file:
+                id_file.write(count_directories(DIRECTORY_PATH))
         
         return links, title
     else:
